src/use-cases/validate-check-in.ts
import { ICheckIn, ICheckInsRepository } from "@/interfaces/check-in";
import { ResourceNotFoundError } from "./errors/resource-not-found-error";
import dayjs from "dayjs";
import { LateCheckInValidationError } from "./errors/late-check-in-validation-error";

interface IValidateCheckInUseCaseRequest {
  checkInId: string;
}

interface IValidateCheckInUseCaseResponse {
  checkIn: ICheckIn;
}

export class ValidateCheckInUseCase {
  constructor(private checkInsRepository: ICheckInsRepository) {}

  async execute({
    checkInId,
  }: IValidateCheckInUseCaseRequest): Promise<IValidateCheckInUseCaseResponse> {
    const checkIn = await this.checkInsRepository.findById(checkInId);

    if (!checkIn) {
      throw new ResourceNotFoundError();
    }

    const distanceInMinutesFromCheckInCreation = dayjs(new Date()).diff(
      checkIn.created_at,
      "minutes"
    );

    const TIME_LIMIT_IN_MINUTES_FOR_CHECKIN_VALIDATION = 20;
    if (
      distanceInMinutesFromCheckInCreation >
      TIME_LIMIT_IN_MINUTES_FOR_CHECKIN_VALIDATION
    ) {
      throw new LateCheckInValidationError();
    }

    checkIn.validated_at = new Date();
    await this.checkInsRepository.update(checkIn);

    return { checkIn };
  }
}
